requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
nltk==3.8.1
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3